Fixed Tests to be independant from one another.
Added the reset function for test purposes (This part should be replace by a Transaction for tests when we migrade from the Hashmap to a database)

Added Integration tests and a few failure responses. (Should maybe be migrated to error 400s, need to discuss about this part)

TODO:
Think of a better way to return responses, these are way too wordy, maybe create decorators for the request and the response that will hold response objects?